backend/requirements.txt
Flask
pymongo
python-dotenv
voxtral
flask_cors
